[dev] -- refactorizando nombres de useToken y parseData, cambiando colores y algunos estilos de la tabla facturas
