+++ eklipse-2.0/src/pages/tableInvoices/table.css
@@ -8,7 +8,7 @@
 
 .table_title {
     align-items: center;
-    background-color: #008080;
+    background-color: #00a6f0;
     color: #ffffff;
     display: flex;
     justify-content: center;
@@ -34,14 +34,14 @@
 .export_invoice--excel {
     background-color: #ffffff;
     border-radius: 10px;
-    color: #008080;
+    color: #00a6f0;
     cursor: pointer;
     height: 80%;
     width: 15%;
 }
 
 .export_invoice--excel:hover {
-    background-color: #008080;
+    background-color: #00a6f0;
     color: #ffffff;
     transition: background 500ms linear;
 }
@@ -53,14 +53,14 @@
 }
 
 .content_table_data {
-    border: 2px solid #008080;
+    border: 2px solid #00a6f0;
     height: 10%;
     text-align: center;
     width: 100%;
 }
 
 .list_invoices {
-    border: 1px solid #008080;
+    border: 1px solid #00a6f0;
     display: flex;
     flex-direction: column;
     height: 80%;
@@ -91,9 +91,8 @@
 
 .invoice--btn_delete {
     align-items: center;
-    background-color: #008080;
+    background-color: #00a6f0;
     cursor: pointer;
-    border-radius: 8px;
     color: #fafbfd;
     display: flex;
     flex-direction: column;
@@ -104,7 +103,7 @@
 
 .download_csv_confirm_modal {
     border: 1px solid #ffffff;
-    background-color: #008080;
+    background-color: #00a6f0;
     display: flex;
     display: none;
     flex-direction: column;
@@ -139,7 +138,7 @@
 }
 
 .btn_download_csv {
-    color: #008080;
+    color: #00a6f0;
     background-color: #ffffff;
     border-radius: 10px;
     height: 40px;
